Multifunctions Class with all the functions
The Python file, `multiFunction.py`, contains a class named `multiFunctions()`. Each function is defined within this class. A separate Jupyter Notebook (`.ipynb`) file was created to call and execute these functions. Critically, the import statement `from multipleFunctions import multipleFunctions` is used to import the associated file and then the class for use in the notebook.